monalisa/requirements.txt
speechrecognition==3.10.0
pyaudio==0.2.14
openai>=1.40.0
httpx>=0.27.0
gtts==2.5.1
opencv-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
mediapipe==0.10.9
pygame==2.5.2
python-dotenv==1.0.1
requests>=2.28.0
librosa==0.10.2
scipy==1.11.4
scikit-image==0.22.0